Move createPncDisposalFromOffence into createDisposalsFromOffence
As createPncDisposalFromOffence is only used for createDisposalsFromOffence.
