+++ packages/core/phase3/lib/hearingDetails/generateHearingsAdjudicationsAndDisposals.ts
@@ -4,7 +4,7 @@ import type { CourtHearingAdjudicationAndDisposal } from "../../types/HearingDet
 import getAdjustedRecordableOffencesForCourtCase from "../../../lib/getAdjustedRecordableOffencesForCourtCase"
 import { createAdjudicationFromOffence } from "./createAdjudicationFromOffence"
 import { createCourtHearingFromOffence } from "./createCourtHearingFromOffence"
-import { createDisposalsFromOffence } from "./createDisposalsFromOffence"
+import createDisposalsFromOffence from "./createDisposalsFromOffence"
 
 export const generateHearingsAdjudicationsAndDisposals = (
   pncUpdateDataset: PncUpdateDataset,
